[Population search] convert index to MUI
